# docs/swagger.json
{
	"openapi": "3.0.0",
	"paths": {
...
		"/api/v1/notifications/channels": {
			"post": {
				"description": "Creates a Telegram or Slack notification channel.",
				"operationId": "NotificationsController_createChannel",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/Function" }
						}
					}
				},
				"responses": {
					"201": { "description": "Channel created successfully" },
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Create notification channel",
				"tags": ["Notifications"]
			},
			"get": {
				"description": "Returns all active notification channels for the user.",
				"operationId": "NotificationsController_listChannels",
				"parameters": [],
				"responses": {
					"200": { "description": "List of channels" },
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "List notification channels",
				"tags": ["Notifications"]
			}
		},
		"/api/v1/notifications/channels/{id}": {
			"delete": {
				"description": "Soft-deletes a notification channel (sets isActive to false).",
				"operationId": "NotificationsController_deleteChannel",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Channel UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Channel deleted successfully" },
					"404": {
						"description": "Not Found - Resource does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Delete notification channel",
				"tags": ["Notifications"]
			}
		},
		"/api/v1/notifications/alert-rules": {
			"post": {
				"description": "Creates an alert rule linking an endpoint to a notification channel.",
				"operationId": "NotificationsController_createAlertRule",
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": { "$ref": "#/components/schemas/Function" }
						}
					}
				},
				"responses": {
					"201": { "description": "Alert rule created successfully" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Create alert rule",
				"tags": ["Notifications"]
			}
		},
		"/api/v1/notifications/alert-rules/endpoint/{endpointId}": {
			"get": {
				"description": "Returns all alert rules for a given endpoint.",
				"operationId": "NotificationsController_getAlertRules",
				"parameters": [
					{
						"name": "endpointId",
						"required": true,
						"in": "path",
						"description": "Endpoint UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "List of alert rules" },
					"403": {
						"description": "Forbidden - Insufficient permissions",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "List alert rules for endpoint",
				"tags": ["Notifications"]
			}
		},
		"/api/v1/notifications/alert-rules/{id}": {
			"delete": {
				"description": "Permanently deletes an alert rule.",
				"operationId": "NotificationsController_deleteAlertRule",
				"parameters": [
					{
						"name": "id",
						"required": true,
						"in": "path",
						"description": "Alert rule UUID",
						"schema": { "format": "uuid", "type": "string" }
					}
				],
				"responses": {
					"200": { "description": "Alert rule deleted successfully" },
					"404": {
						"description": "Not Found - Resource does not exist",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"429": {
						"description": "Too Many Requests - Rate limit exceeded",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					},
					"500": {
						"description": "Internal Server Error - Unexpected server error",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/ErrorResponseSchema" }
							}
						}
					}
				},
				"security": [{ "Bearer": [] }],
				"summary": "Delete alert rule",
				"tags": ["Notifications"]
			}
		}
	},
	"info": {
		"title": "Proxy Server API",
		"description": "API documentation for the Proxy Server backend",
		"version": "1.0",
		"contact": { "name": "Proxy Server", "url": "", "email": "" },
		"license": { "name": "ISC", "url": "" }
	},
...
	"servers": [
		{ "url": "http://localhost:3000", "description": "development" },
		{ "url": "http://lvh.me:3000", "description": "production" }
	],
	"components": {
		"securitySchemes": {
			"Bearer": {
				"scheme": "bearer",
				"bearerFormat": "JWT",
				"description": "Enter JWT access token from /auth/sign-in, /auth/verify-email, or /auth/refresh",
				"in": "header",
				"name": "Authorization",
				"type": "http"
			}
		},
		"schemas": {
...
			"Function": { "type": "object", "properties": {} },
			"ErrorResponseSchema": {
				"type": "object",
				"properties": {
					"statusCode": {
						"type": "number",
						"description": "HTTP status code",
						"example": 400,
						"minimum": 400,
						"maximum": 599
					},
					"error": {
						"type": "string",
						"description": "Error type or exception name",
						"example": "BadRequestException"
					},
					"message": {
						"description": "Human-readable error message(s)",
						"oneOf": [
							{ "type": "string", "example": "Validation failed" },
							{
								"type": "array",
								"items": { "type": "string" },
								"example": [
									"email must be an email",
									"password must be longer than 8 characters"
								]
							}
						]
					},
					"path": {
						"type": "string",
						"description": "Request path that triggered the error",
						"example": "/api/v1/auth/register"
					},
					"timestamp": {
						"type": "string",
						"description": "ISO 8601 timestamp when the error occurred",
						"example": "2025-03-18T12:00:00.000Z"
					}
				},
				"required": ["statusCode", "error", "message"]
			}
		}
	},
	"security": [{ "Bearer": [] }]
}